I remove repetitive classes and resume in category file
