+++ lib/bot.rb
@@ -1,12 +1,7 @@
 # rubocop: disable Metrics/BlockLength,Layout/LineLength
 
 require 'telegram/bot'
-require_relative 'training'
-require_relative 'study'
-require_relative 'work'
-require_relative 'relax'
-require_relative 'dance'
-require_relative 'romance'
+require_relative 'category'
 
 class Bot
   token = '918223094:AAFwkiG1yn4Vsqg_Y-aO4lIZOwx6e8Ava7g'
@@ -17,28 +12,28 @@ class Bot
       when '/start'
         bot.api.send_message(chat_id: message.chat.id, text: "Hi, #{message.from.first_name}, Welcome to music motivation, this bot will give you random youtube links of music depending of the category that you choose. Use /start to start the bot, /stop to stop the bot, You can choose beetwing this categories just write any of this: /training, /study, /work, /relax, /dance, /romance")
       when '/training'
-        mood = Training.new
-        song = mood.list_pick
+        mood = Category.new
+        song = mood.list_pick(mood.training)
         bot.api.send_message(chat_id: message.chat.id, text: song.to_s, date: message.date)
       when '/study'
-        mood = Study.new
-        song = mood.list_pick
+        mood = Category.new
+        song = mood.list_pick(mood.study)
         bot.api.send_message(chat_id: message.chat.id, text: song.to_s, date: message.date)
       when '/work'
-        mood = Work.new
-        song = mood.list_pick
+        mood = Category.new
+        song = mood.list_pick(mood.work)
         bot.api.send_message(chat_id: message.chat.id, text: song.to_s, date: message.date)
       when '/relax'
-        mood = Relax.new
-        song = mood.list_pick
+        mood = Category.new
+        song = mood.list_pick(mood.relax)
         bot.api.send_message(chat_id: message.chat.id, text: song.to_s, date: message.date)
       when '/dance'
-        mood = Dance.new
-        song = mood.list_pick
+        mood = Category.new
+        song = mood.list_pick(mood.dance)
         bot.api.send_message(chat_id: message.chat.id, text: song.to_s, date: message.date)
       when '/romance'
-        mood = Romance.new
-        song = mood.list_pick
+        mood = Category.new
+        song = mood.list_pick(mood.romance)
         bot.api.send_message(chat_id: message.chat.id, text: song.to_s, date: message.date)
       when '/stop'
         bot.api.send_message(chat_id: message.chat.id, text: "Bye, #{message.from.first_name}", date: message.date)
